+++ spimdisasm/mips/symbols/MipsSymbolBase.py
@@ -301,7 +301,7 @@ class SymbolBase(common.ElementBase):
         comment = self.generateAsmLineComment(localOffset+j)
         return f"{comment} {dotType} {value}"
 
-    def getNthWordAsBytesAndShorts(self, i: int, sym1: common.ContextSymbol|None, sym2: common.ContextSymbol|None, sym3: common.ContextSymbol|None, lastSymName: str) -> tuple[str, int]:
+    def getNthWordAsBytesAndShorts(self, i: int, sym1: common.ContextSymbol|None, sym2: common.ContextSymbol|None, sym3: common.ContextSymbol|None, lastSymName: str, symSize: int | None) -> tuple[str, int]:
         output = ""
 
         # Check the 4 bytes of this word to determine if each pair of bytes should be disassembled as `.short`s or a pair of `.byte`s
@@ -316,9 +316,10 @@ class SymbolBase(common.ElementBase):
             output += self.getJByteAsByte(i, 0)
             output += common.GlobalConfig.LINE_ENDS
 
-            if sym1 is not None:
+            if sym1 is not None or i * 4 + 1 == symSize:
                 output += self._getSymEnd(lastSymName)
-                lastSymName = sym1.getName()
+                if sym1 is not None:
+                    lastSymName = sym1.getName()
 
             output += self.getExtraLabelFromSymbol(sym1)
             output += self.getJByteAsByte(i, 1)
@@ -327,9 +328,10 @@ class SymbolBase(common.ElementBase):
             output += self.getJByteAsShort(i, 0)
             output += common.GlobalConfig.LINE_ENDS
 
-        if sym2 is not None:
+        if sym2 is not None or i * 4 + 2 == symSize:
             output += self._getSymEnd(lastSymName)
-            lastSymName = sym2.getName()
+            if sym2 is not None:
+                lastSymName = sym2.getName()
 
         output += self.getExtraLabelFromSymbol(sym2)
         if sym3 is not None or (sym2 is not None and sym2.isByte()) or (self.isByte(i) and (sym2 is None or not sym2.isShort())):
@@ -342,9 +344,10 @@ class SymbolBase(common.ElementBase):
             output += self.getJByteAsByte(i, 2)
             output += common.GlobalConfig.LINE_ENDS
 
-            if sym3 is not None:
+            if sym3 is not None or i * 4 + 3 == symSize:
                 output += self._getSymEnd(lastSymName)
-                lastSymName = sym3.getName()
+                if sym3 is not None:
+                    lastSymName = sym3.getName()
 
             output += self.getExtraLabelFromSymbol(sym3)
             output += self.getJByteAsByte(i, 3)
@@ -661,7 +664,7 @@ class SymbolBase(common.ElementBase):
         canReferenceSymbolsWithAddends = self.canUseAddendsOnData()
         canReferenceConstants = self.canUseConstantsOnData()
 
-        symSize: int | None = self.contextSym.getSize()
+        symSize: int | None = self.contextSym.userDeclaredSize
 
         i = 0
         while i < self.sizew:
@@ -674,7 +677,7 @@ class SymbolBase(common.ElementBase):
 
             # Check for symbols in the middle of this word
             if sym1 is not None or sym2 is not None or sym3 is not None or self.isByte(i) or self.isShort(i):
-                data, skip = self.getNthWordAsBytesAndShorts(i, sym1, sym2, sym3, lastSymName)
+                data, skip = self.getNthWordAsBytesAndShorts(i, sym1, sym2, sym3, lastSymName, symSize)
 
                 if sym3 is not None:
                     lastSymName = sym3.getName()
@@ -717,9 +720,11 @@ class SymbolBase(common.ElementBase):
             if i * 4 == symSize:
                 output += self._getSymEnd(lastSymName)
 
-        if symSize is None:
+        if symSize is None or i * 4 < symSize:
             # Size was cleared up by some in-middle symbol, so we need to emit
-            # the end label for the last sym
+            # the end label for the last sym.
+            # Or the user specified a larger size than what the symbol symbol
+            # actually is.
             output += self._getSymEnd(lastSymName)
 
         nameEnd = self.getNameEnd()